requirements.txt
gym-chess
python-chess
torch
stable-baselines3[extra]
numpy
tqdm
zstandard
